Add basic support for i.MX6UL
Backport code from Freescales U-Boot for i.MX6UL which is based on
u-boot-2015.04. The code is taken from rel_imx_3.14.38_6ul_ga (in
branch imx6_v2015.04_3.14.38_6ul_ga) of GIT repository

  git://git.freescale.com/imx/uboot-imx.git

We have only taken code that has direct impact on i.MX6UL. Code for
other CPUs (e.g. i.MX7D, i.MX6QP) was left out.

- Add CONFIG_MX6UL to define an i.MX6UL architecture.
- Modify register and clock settings to support i.MX6UL.
- Add pin configuration macros for i.MX6UL.
- Add second argument for enable_fec_anatop_clock() to allow more
  than one ethernet port.
- Add UL specific code in drivers/net/fec_mxc.c

+++ arch/arm/include/asm/arch-mx6/sys_proto.h
@@ -2,6 +2,8 @@
  * (C) Copyright 2009
  * Stefano Babic, DENX Software Engineering, sbabic@denx.de.
  *
+ * (C) Copyright 2009-2015 Freescale Semiconductor, Inc.
+ *
  * SPDX-License-Identifier:	GPL-2.0+
  */
 
@@ -12,7 +14,7 @@
 #include "../arch-imx/cpu.h"
 
 #define soc_rev() (get_cpu_rev() & 0xFF)
-#define is_soc_rev(rev)        (soc_rev() - rev)
+#define is_soc_rev(rev)        (int)(soc_rev() - rev)
 
 u32 get_nr_cpus(void);
 u32 get_cpu_rev(void);
